[webview_flutter] Remove non-public API
ewk_settings_viewport_meta_tag_set() was moved to ewk_settings_product.h in tizen 6.5.
Since this header is non-public, we can no longer use this api.
By default, the viewport meta tag is enabled on mobile and wearable, but it is disabled on TV.
Since webview_flutter_tizen only support the TV profile, it doesn't matter if we don't use this API.