[webview_flutter_lwe] Supports multiple JavaScriptChannel method call
Previously, javascript channel methods that are called
before the webview is created are stored in the pending list.(based on name)
Because of that, there was a problem with channels being missing
when adding multiple channels.

When the javascript channel method is called,
it is stored differently depending on the name to prevent it from being missed.

```dart
controller
..addJavaScriptChannel(
        'method1',...)
..addJavaScriptChannel(
        'method2',...);
```