[video_player] Live streaming content starts playing immediately when SeekTo() is called
According to Tizen media player's member, Tizen player does not support seekTo for live streaming content.

Therefore, duration is returned as 0. When duration is 0, an error occurs due to a zero divide error(NaN value) in the progress bar of video_player.
And the resume behavior(play()) of video_player calls seekTo(0) when the duration and position are the same(when the video has been played to the end), which causes the video to not be played again in the case of live streaming content.

Therefore, in the case of live, if set_play_position fails, call play() immediately to playback.
A