[audioplayers] Fix play() didn't work after setSource()
When setSource() is called, or setSource() is called when play() is called,
it waits for the event of AudioEventType.prepared. If the url has not been set
or a different url is set than before, it calls the prepared event, but if the url is the same,
it does not call the event and waits for preparedCompeter.
Therefore, regardless of whether the url is set, it always resets the player and calls preparePlayer().