[audioplayers] Changed to create a player when AudioPlayer created
Starting with audioplayers 6.1.0, PositionUpdater.stopAndUpdate() is now called when AudioPlayer.stop() is called.
Previously, since play() had to be called to create a player,
calling stop() first would throw an Exception when calling getPosition()(returned an invalid parameter).
Therefore, to prevent this, a player is also created when an AudioPlayer object is created.