[audioplayers] Skip exception when getPosition is called after release in release mode
When stop() or pause() is called in AudioPlayer 6.1.0,
PositionUpdater's stopAndUpdate() is called. At this time,
getPosition() is called, but in ReleaseMode, the player is released after Stop(),
so an Exception is thrown. Since there are differences from the implementation
in the Frontend package, an Exception is not thrown in this case.