[video_player_avplay] Automatically rotates video player based on device orientation
We already provide rotate API for video player,
but it seems like a bug that the video player's video image doesn't rotate according to the device orientation.
So we should automatically change this inside video_player.
When user call the API, the video image may rotate,
but when the device's orientation changes, the rotation may change again.